cms/requirements.txt
asgiref==3.3.1
coverage==5.5
Django==3.1.7
django-phonenumber-field==5.0.0
django-seed==0.2.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Faker==6.6.0
mysqlclient==2.0.3
phonenumbers==8.12.19
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
